Move core_extensions under lib, and move ActiveRecord monkeypatches under core_extensions
