+++ lib/active_record/connection_adapters/clickhouse/schema_creation.rb
@@ -125,7 +125,7 @@ module ActiveRecord
         end
 
         def current_database
-          if ActiveRecord::VERSION::MAJOR < 6
+          if ActiveRecord::version < Gem::Version.new('6.0.0')
             ActiveRecord::Base.connection_config[:database]
           else
             ActiveRecord::Base.connection_db_config.database
