+++ lib/core_extensions/active_record/migration/command_recorder.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module CoreExtensions
   module ActiveRecord
     module Migration
