+++ spec/cases/multi_db_compatibility_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-if ActiveRecord::VERSION::MAJOR >= 6
+if ActiveRecord::version >= Gem::Version.new('6.0.0')
   RSpec.describe 'multi-database compatibility', :migrations do
     let(:model) do
       Class.new(ActiveRecord::Base) do
